requirements.txt
Django==4.2.*
Pillow==9.5.*
django-admin-sortable2==2.1.*
django-tinymce==3.6.*
